# Performance improvements, added apiCalming.js