--- a/src/App.jsx
+++ b/src/App.jsx
@@ -14,6 +14,7 @@ import initializejQuery from './js/utils/initializejQuery';
 import { renderLog } from './js/utils/logging';
 import RouterV5SendMatch from './js/utils/RouterV5SendMatch';
 // import SnackNotifier from './js/components/Widgets/SnackNotifier';
+import { numberOfNeedlesFoundInString } from './js/utils/searchFunctions';
 
 // Root URL pages
 
@@ -139,6 +140,9 @@ class App extends Component {
     // this.InitializeOnce();
     console.log('href in App.js componentDidMount: ', window.location.href);
 
+    // Aug 2019: Adding functions to the String prototype will make stuff like `for (char in str)` break, because it will loop over the substringOccurences property.
+    // As long as we use `forEach()` or `for (char of str)` then that side effect will be mitigated.
+    String.prototype.numberOfNeedlesFoundInString = numberOfNeedlesFoundInString; // eslint-disable-line
   }
 
   componentDidCatch (error, info) {
@@ -189,6 +193,7 @@ class App extends Component {
             <ThemeProvider theme={styledTheme}>
               <WeVoteRouter>
                 <WeVoteBody>
+                  {/* DO NOT put SnackNotifier or anything else that is non-essential here (to keep it out of the main chuck. */}
                   <Header params={{ }} pathname={window.location.href} />
                   <Switch>
                     <Route exact path="/about"><About /></Route>