--- a/src/js/actions/VoterActions.js
+++ b/src/js/actions/VoterActions.js
@@ -261,6 +261,7 @@ export default {
   },
 
   voterRetrieve () {
+    // console.log('VoterActions, voterRetrieve');
     Dispatcher.loadEndpoint('voterRetrieve');
   },