--- a/src/js/compileDate.js
+++ b/src/js/compileDate.js
@@ -1,1 +1,1 @@
-module.exports = ['5/14/2021, 3:44:21 PM'];
+module.exports = ['5/19/2021, 10:49:20 AM'];