--- a/src/js/compileDate.js
+++ b/src/js/compileDate.js
@@ -1,1 +1,1 @@
-module.exports = ['5/19/2021, 10:49:20 AM'];
+module.exports = ['5/20/2021, 5:43:15 PM'];