--- a/src/js/compileDate.js
+++ b/src/js/compileDate.js
@@ -1,1 +1,1 @@
-module.exports = ['5/20/2021, 5:43:15 PM'];
+module.exports = ['5/24/2021, 9:50:23 AM'];