--- a/src/js/compileDate.js
+++ b/src/js/compileDate.js
@@ -1,1 +1,1 @@
-module.exports = ['5/24/2021, 9:50:23 AM'];
+module.exports = ['5/25/2021, 1:36:41 PM'];