--- a/src/js/components/Activity/ActivityPostAdd.jsx
+++ b/src/js/components/Activity/ActivityPostAdd.jsx
@@ -3,7 +3,6 @@ import { withStyles, withTheme } from '@material-ui/core/styles';
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import styled from 'styled-components';
-import stockAvatar from '../../../img/global/icons/avatar-generic.png';
 import VoterStore from '../../stores/VoterStore';
 import { cordovaNewsPaddingTop } from '../../utils/cordovaOffsets';
 import { cordovaDot, isCordova } from '../../utils/cordovaUtils';
@@ -11,6 +10,7 @@ import { renderLog } from '../../utils/logging';
 
 const ActivityPostModal = React.lazy(() => import(/* webpackChunkName: 'ActivityPostModal' */ './ActivityPostModal'));
 
+const stockAvatar = '../../../img/global/icons/avatar-generic.png';
 
 
 class ActivityPostAdd extends Component {