… package.json
… src/css/main.css
… src/index.html
… src/js/components/Navigation/HeaderNotificationMenu.jsx
… src/js/components/OpinionsAndBallotItems/CandidateItemForOpinions.jsx
… src/js/components/OpinionsAndBallotItems/VoterGuideDisplayForListForOpinions.jsx
… src/js/components/Organization/OrganizationDisplayForList.jsx
… src/js/components/Organization/OrganizationPopoverCard.jsx
… src/js/components/Search/SearchBar.jsx
… src/js/components/Settings/DragAndDrop.jsx
… src/js/components/Settings/PaidAccountUpgradeModal.jsx
… src/js/components/Settings/SettingsAccount.jsx
… src/js/components/Settings/SettingsProfilePicture.jsx
… src/js/components/Settings/SettingsSubscriptionPlan.jsx
… src/js/components/Settings/SettingsWidgetFirstLastName.jsx
… src/js/components/Twitter/TwitterAccountCard.jsx
… src/js/components/Twitter/TwitterSignIn.jsx
… src/js/components/Values/IssueImageDisplay.jsx
… src/js/components/Values/ValuesToFollowPreview.jsx
… src/js/components/VoterGuide/OrganizationCard.jsx
… src/js/components/VoterGuide/OrganizationVoterGuideCandidateItem.jsx
… src/js/components/VoterGuide/OrganizationVoterGuideCard.jsx
… src/js/components/VoterGuide/VoterGuideDisplayForList.jsx
… src/js/components/VoterGuide/VoterGuidePositionItem.jsx
… src/js/components/Welcome/WelcomeFooter.jsx
… src/js/components/Widgets/EndorsementCard.jsx
… src/js/components/Widgets/FirstAndLastNameRequiredAlert.jsx
… src/js/components/Widgets/FollowToggle.jsx
… src/js/components/Widgets/PositionStatementModal.jsx
… src/js/components/Widgets/ThisIsMeAction.jsx